Logging all Serial messages to a persistent store, always in AP mode giving access to log, no longer forgetting WIFI on unsuccessful connections.
The forgetting WIFI mode was a bad idea because for example short router downtime occurs, the device enters power-hungry mode and drains battery.

Storing messages to persistent store is useful for serving them to a local Android device. In the future, a command will upload them to S3.

Refactoring some code to make it better look like a state machine (thus easier to modify and more scalable).