Failed attempt to migrate from MQTT to REST AWS IoT.
Motivation was that a bridge is currently needed because ESP does not
support authentication with certs. However, turns out SSL requests do consume a lot of
memory (both for creating the V4 key and for sending the request itself). This is why
I could not get it to work (got as far as buidling the signed request and sending it, but sending fails).

Future attempts may try to cut down on memory usage elsewhere, for example starting from scratch and
adding functionality gradually.