Removing the process_gpio state, wifi and mqtt connection speed optimizaitons. Flipping high and low actions.
process_gpio is replaced by actions now.

Removing wifi and mqtt connection optimizations
to provide a resilient connectivity for always-on devices. The device would not correctly
understand it's connectivity state and stop publishing, although working fine.

There was a related bug which did not reset wifi connectivity counter on boot, so this may be the reason for the
connectivity misunderstanding. However, the optimizations + the bug fix were not tested.

Flipping the HIGH and LOW states was useful because on device being off, the pin will be low,
and low closing the windows is the desired behavior.