Working expectation and ssd included in sense value.
The expectation and ssd are calculated using a modified Welford algorithm
which uses with the N latest datapoints for a sense (currently set to 10).

A couple of questions of the validity of the current implementation - first, take care and follow the TODO
by removing sense values present in last reading but non-present in this reading.

Then, do we take into account "hardware"-wrong values (like checksum failures) in the expectation? Probably no,
so do the corrections.

Finally, variance is currently set to start at 10 because at 0 it finds it hard to move from 0 (due to numerical rounding, using integers while the increments are doubles. Think whether this is what you want and observe behaviour.